Blacklist functionality
A user is blacklisted if it exceeds the processing limit stated. The user is detected with his IP address and user agent string. If any of both exists in any register of the database those are updated in the connection attempts as well as the date attempted except when it has exceeded the limit.